% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatPairs-data.R
\docType{data}
\name{CatPairs}
\alias{CatPairs}
\title{Categorical Cause-Effect Pairs}
\format{
A list of length 2. The first element is a list of 33 cause-effect pairs as data frames with the first column being the cause and the second column being the effect. The second element is a list of sources of each pair.
}
\usage{
data(CatPairs)
}
\description{
Cause-effect pairs extracted from R packages MASS and datasets for which the pairwise causal relationships are clear from the context, and at least one of the variables in each pair is categorical. For non-categorical variable, we discretized it at 5 evenly spaced quantiles.The current version contains 33 categorical cause-effect pairs.
}
\keyword{datasets}
